commit 3226dfc1f52098d99aaafd4dda78064f924db8d2
Author: Daniel Collin <daniel@collin.com>
Date:   Mon Aug 31 19:07:32 2020 +0200

    Bump to 1.0.0

commit a3f05c0bfdeaf1e5e11a8f0f92174323fc283786
Author: Daniel Collin <daniel@collin.com>
Date:   Mon Aug 31 19:06:25 2020 +0200

    Bump sid buffer

commit 8f2151fe86bf17086b10a4c415fa54aff4f483f7
Author: Daniel Collin <daniel@collin.com>
Date:   Mon Aug 31 19:05:40 2020 +0200

    Upgraded to latest libresidfp-svn

    Fixes #167

commit dde9b4a1ccf97b761a0348c4a9ceb8fb59be8ecd
Author: Daniel Collin <daniel@collin.com>
Date:   Sun Aug 30 10:23:02 2020 +0200

    Use templates for bugs/requests

+++ src/plugins/playback/sid/libsidplayfp/src/builders/residfp-builder/residfp/Filter.h
@@ -127,7 +127,7 @@ public:
      * @param v3 voice 3 in
      * @return filtered output
      */
-    virtual int clock(int v1, int v2, int v3) = 0;
+    virtual unsigned short clock(int v1, int v2, int v3) = 0;
 
     /**
      * Enable filter.
